+++ servente/src/main.rs
@@ -12,6 +12,7 @@ use tokio::task;
 
 use std::{io, sync::Arc, time::Instant, env::current_dir};
 
+mod abnf;
 mod cert;
 mod client;
 mod example_handlers;